Add files via upload
Question 1
The number, 197, is called a circular prime because all rotations of the digits: 197, 971, and 719, are themselves prime.
There are thirteen such primes below 100: 2, 3, 5, 7, 11, 13, 17, 31, 37, 71, 73, 79, and 97.
How many circular primes are there below One million?
Output:
25

Question 2
The prime 41, can be written as the sum of six consecutive primes:
41 = 2 + 3 + 5 + 7 + 11 + 13
This is the longest sum of consecutive primes that adds to a prime below one-hundred. The longest sum of consecutive
primes below one-thousand that adds to a prime, contains 21 terms, and is equal to 953.
Which prime, below one-million, can be written as the sum of the most consecutive primes?
Output:
78,498

Question 3
The number 3797 has an interesting property. Being prime itself, it is possible to continuously remove digits from left to right,
and remain prime at each stage: 3797, 797, 97, and 7. Similarly we can work from right to left: 3797, 379, 37, and 3.
Find the sum of the only eleven primes that are both truncatable from left to right and right to left.
NOTE: 2, 3, 5, and 7 are not considered to be truncatable primes.
Output:
23
37
53
73
313
317
373
797
3137
3797
739397
R e s ul t i s : 748317

Question 4
If we take 47, reverse and add, 47 + 74 = 121, which is palindromic.
Not all numbers produce palindromes so quickly. For example,
349 + 943 = 1292,
1292 + 2921 = 4213
4213 + 3124 = 7337
1. 56 becomes palindromic after one iteration: 56+65 = 121.
2. 57 becomes palindromic after two iterations: 57+75 = 132, 132+231 = 363.
Last Date: Monday, November 22, 2021 11:59:00 PM 1
Assignment#1 Programming for Big Data (PBD)
3. 59 becomes a palindrome after 3 iterations: 59+95 = 154, 154+451 = 605, 605+506 = 1111
4. 89 takes an unusually large 24 iterations (the most of any number under 10,000 that is known to resolve into a palindrome)
to reach the palindrome 8,813,200,023,188.
5. 10,911 reaches the palindrome 4668731596684224866951378664 (28 digits) after 55 steps.
6. 1,186,060,307,891,929,990 takes 261 iterations to reach the 119-digit palindrome 44562665878976437622437848976653870388884783662598425855963436955852489526638748888307835667984873422673467987856626544,
which is the current world record for the Most Delayed Palindromic Number
Although no one has proved it yet, it is thought that some numbers, like 196, never produce a palindrome. A number that
never forms a palindrome through the reverse and add process is called a Lychrel number. Due to the theoretical nature of these
numbers, and for the purpose of this problem, we shall assume that a number is Lychrel until proven otherwise. In addition you
are given that for every number below ten-thousand, it will either (i) become a palindrome in less than fifty iterations, or, (ii) no
one, with all the computing power that exists, has managed so far to map it to a palindrome. In fact, 10677 is the first number
to be shown to require over fifty iterations before producing a palindrome: 4668731596684224866951378664 (53 iterations,
28-digits).
Surprisingly, there are palindromic numbers that are themselves Lychrel numbers; the first example is 4994.
How many Lychrel numbers are there below ten-thousand?
Output:
249

Question 5
Find the minimum value in the circular queue without using any extra space. Data in Queue should be intact. Following are the
only functions that should be used to write this program: Enqueue | Dequeue | Getfront_value | isEmpty | isFull.